agents_llms/pyproject.toml
[tool.poetry]
name = "agents-llms"
version = "0.1.0"
description = ""
authors = ["Vladimir Tamayo <vladimirtamayo@edrolo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
gradio = "^5.20.1"
openai = "^1.66.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
